fix: improve Redshift AWS credential chain handling
- Add loadAWSCredentialsFromChain helper for loading credentials from
  the default AWS chain (env vars, profiles, IAM roles)
- Handle errors from ensureAWSCredentials in getS3Props with warning
  instead of silently ignoring
- Add redactCredentials method to mask AWS secrets for safe logging
- Fix USE_ENVIRONMENT check to properly handle boolean conversion
  instead of only checking for "false" string
- Clarify documentation and error messages around credential fallback
